requirements.txt
openai
pinecone
python-dotenv
tiktoken
